datastore: introduce a Transaction type to encapsulate transactions.
Currently, we don't support mutations correctly during a transaction, so
that the first mutation (Put/Delete) causes the transaction to be
committed, and any later mutations to fail. It also causes spurious
failures when the transaction scope ends.

Instead, we introduce a new type Transaction which has methods for
Put/Delete/Get and can be used to build a Query. Any mutations will
be applied at once during the Commit method call.

Change-Id: Icac1f0ab2261ac52f755eea11948f4cca43bb2e6