cloud/storage: Consistently request the "full" projection.
This isolates us from the crazy default behavior of the wrapped
storage/v1 package, which requests different projections for different
methods, sometimes conditioned on input.

In particular, it's a problem for code I'm writing that ACLs currently
show up in the result of UpdateAttrs, but not in the result of
ListObjects. Since there is no mechanism offered for configuring this as
an option, the only sane behavior is to always return them.

Change-Id: I210e0cd281c07ad4dfaa66bbff6c014a4f5efda5