cloud/pubsub: Changed to use base64.StdEncoding.DecodeString.
This CL changes which encoding to use for encode/decode. It turned out
that Cloud Pub/Sub server's frontend use StdEncoding to encode the
message body, and we must use StdEncoding at least for decoding the
messages. Also, the server accepts both encodings, so I also use
StdEncoding for encoding the message body.

Change-Id: I35b616d42ac3feb86c858fb3b203b00977db4587