pubsub: Add FastMode for prompt acker shutdown.
Once Iterator.Stop is called, messages are now acked opportunistically
as they arrive in the acker. This ensures that Iterator.Stop returns
promptly once the final Message is acked.

Note: Messages are still acked in batches, and only one ack request will
be in flight at once, but any pending messages will be acked as soon as
the previous ack request has returned.

The introduction of ackBuffer also sets the stage for more flexibile
decisions about when to ack, e.g. in the future we may send an ack
request when the buffer reaches a certain size.

Fixes #236.

Change-Id: I2dfd9ac39a5abc3d85e76dc04d0320428222cade
Reviewed-on: https://code-review.googlesource.com/4431
Reviewed-by: Dave Day <djd@golang.org>