storage: ObjectIterator
Write a standard (or as nearly as possible) iterator for objects in a
bucket.

The underlying method returns pages of object attributes. When
delimiter is set, each page also includes a list of prefixes (which can
be thought of as immediate subdirectories). Since the underlying
iterator is in effect traversing two sequences at once, we can't easy
fit it into item-at-a-time iteration. So we disallow the use of a
delimiter when calling Next.

Deprecate BucketHandle.List and Query.MaxResults.

Change-Id: I200007b92f0fa8ac3691db7c9e0a215f04bfe073
Reviewed-on: https://code-review.googlesource.com/5153
Reviewed-by: Dave Day <djd@golang.org>