preview/logging: deflake iteration
When we use Log, it bundles the entries so that they end up getting sent
to the server at the same time. Some may then get the same timestamp,
and perhaps the insert IDs are random. The service's list method sorts
by first timestamp and then insert ID, so the entries can come back in
a different order than they were passed to Log.

So for repeatability, we set the insert ID.

Change-Id: I0781c85761add26f4c6ec9c0659ab0e26dcbc9e1
Reviewed-on: https://code-review.googlesource.com/7330
Reviewed-by: Chris Broadfoot <cbro@google.com>
Reviewed-by: Dave Day <djd@golang.org>