preview/logging: remove sink output format
The Stackdriver Logging team does not want to expose V1 output format at
all. They want everything to be V2.

New sinks will all have the V2 format.

It won't be possible to learn the output format of existing sinks.
If this ever becomes an issue, we can add it back.

Change-Id: I41e3986e8c7a7797bd37d2365d56fc50faec4b42
Reviewed-on: https://code-review.googlesource.com/7436
Reviewed-by: Brad Fitzpatrick <bradfitz@golang.org>