preview/logging: replace OrderBy with NewestFirst
Remove the OrderBy option for listing log entries. Add NewestFirst.

The logging API is unlikely to have any OrderBy values other than
"timestamp asc" (the default) and "timestamp desc". So for a simpler,
less error-prone API, we remove the general OrderBy option and add
NewestFirst, which is equivalent to "timestamp desc".

Change-Id: I4a02008b04ecdf2ca1e2c1c6c62880505a2990b7
Reviewed-on: https://code-review.googlesource.com/7440
Reviewed-by: Brad Fitzpatrick <bradfitz@golang.org>