datastore: correctly expose ErrInvalidEntityType
The putMutations func was previously wrapping any errors found when
converting an entity using fmt.Errorf, rather than the intended
MultiError.

By wrapping with MultiError, the single-version Put methods correctly
unwrap the original error and so return ErrInvalidEntityType as
expected. It also means that the PutMulti methods are consistent with
the LoadMulti/DeleteMulti methods and the errors are correctly
correlated to the incoming entities.

Fixes #364

Also adds the stub to allow simply faking out the datastore service, to
avoid depending on the production service where possible. I use this for
the simple BeginTransaction RPC in testing this change.

Issue #371

Change-Id: Iac8af341e2df80e8431ef2498287fc9c9a1ab742
Reviewed-on: https://code-review.googlesource.com/7590
Reviewed-by: Michael McGreevy <mcgreevy@golang.org>