pubsub: add Close method
Every client should have a Close method, even if it's a no-op.

That way we can backwards-compatibly add closeable resources to the
client.

In anticipation of the time (in a couple of months) when pubsub will
move on top of gRPC (which does not pool connections), add a comment
to the Client type suggesting reuse.

Clarify that it's okay to skip calling Close if you're going to
terminate.

Change-Id: Id965bf2f9ae6388286e71cf3cac43dcf44c51944
Reviewed-on: https://code-review.googlesource.com/7693
Reviewed-by: Ross Light <light@google.com>