storage: add CopierFrom and ComposerFrom
Copying is handled via a Copier object, which provides for setting
destination attributes and accessing the rewrite token.

The API's copy and rewrite operations are collapsed into
Copier.Run. It always calls rewrite. We lose no functionality or
performance in doing so, because the storage service's copy calls
rewrite.

Multiple calls to rewrite are handled automatically. The user can
provide a callback to get progress on the number of bytes transferred,
and to stop the copy.

For consistency, there is also an ObjectHandle.ComposeFrom method and a
Composer type.

ObjectHandle.CopyTo and ObjectHandle.ComposeFrom are deprecated.

This is not a breaking change.

Change-Id: I2f3fa666af47867da0478efd2d8d872674586074
Reviewed-on: https://code-review.googlesource.com/7697
Reviewed-by: Dave Day <djd@golang.org>