Error on res.Row=nil from ApplyReadModifyWrite to avoid panic
Sometimes when ApplyReadModifyWrite is called it will return a
response whose res.Row is nil. This doesn't happen frequently
but it does happen with some regularity. I believe it's due to
a Bigtable error whose details are being lost in the empty response.
From there GRPC + ProtoBufs do their thing and end up returning
res.Row=nil. The problem lies in the caller's assumption that if
err != nil then res.Row will not be nil, but that isn't the case
in reality and a panic ensues.

This fix changes the caller's assumptions but does not attempt to
identify the underlying error from Bigtable.

Change-Id: I4b8e3e25b96b9fd29780862fa05066f70ea4e1bc
Reviewed-on: https://code-review.googlesource.com/9190
Reviewed-by: Gary Elliott <garyelliott@google.com>