bigquery: use int64 for Integer columns
BREAKING CHANGE. No API change, but clients who expect to get an int for
an Integer column will break at runtime.

Prior to this CL, the value from an Integer column was store in an int.
This could result in errors when reading large values on 32-bit
machines. This CL uses int64 instead.

Change-Id: I6b4b4f1106c669190df05bb91031cd53080ee18a
Reviewed-on: https://code-review.googlesource.com/9331
Reviewed-by: Ross Light <light@google.com>