bigtable: Fix cbt required flag enforcement
Currently we reject several valid cbt command invocations, such as 'cbt
help', or 'cbt -project foo listinstances', because the code blindly
enforces the presence of both -project and -instance flags for
everything.

This change instead annotates the commandlist with a set of the required
flags, and the code now checks for the required flags for each command
before invoking it.

I tested this by invoking the above as well as some commands that
require both an instance and project, and verified that everything
behaved as expected.

I also updated scantest and loadtest, will figure out how to run those
tests as well prior to submitting.

Change-Id: I66a8359221fe6936752308dc7e41765bfc57a633
Reviewed-on: https://code-review.googlesource.com/9433
Reviewed-by: Gary Elliott <garyelliott@google.com>