fields: copy FieldsByNameFunc from reflect
This CL is the first in a series whose final state will match
https://code-review.googlesource.com/#/c/9330. The end goal is a
function that will return all the fields of a struct type that are
eligble for serialization, using the same rules as encoding/json.

This CL is useless on its own. Its only purpose is to establish that
the initial logic is taken directly from a function in the standard
library.

PS1 of this CL a copy of the FieldByNameFunc method from the reflect package.
See https://raw.githubusercontent.com/golang/go/release-branch.go1.7/src/reflect/type.go
for the original.

PS2 an almost identical version of that function, modified only to use
exported symbols of the reflect package.

Change-Id: If824825b2921d9cb9eed15b9dfc92da480784cf0
Reviewed-on: https://code-review.googlesource.com/9472
Reviewed-by: Michael Darakananda <pongad@google.com>
Reviewed-by: Jun Mukai <mukai@google.com>