fields: initial support for tags
This CL is a step towards https://code-review.googlesource.com/#/c/9330.

It is the first CL that pays attention to tags.

To keep it simple, we assume the tag name is always "test", and its value is just a name.
Future CLs will expand on that.

We also don't implement all of encoding/json's rules about tags, yet.

We have to change from using reflect.StructField to our own Field
type, so we can record whether the name came rfrom a tag. This is
going to matter in the next CL.

Change-Id: I35f9f013c3abced84f0c2e3e1034c223b8547ffc
Reviewed-on: https://code-review.googlesource.com/9478
Reviewed-by: Jun Mukai <mukai@google.com>