bigtable/cmd/cbt: allow overriding TLS certificate chain
In order to run cbt in some test environments, we need to explicitly
specify the TLS certificates to use. This change allows us to do that
with the new -cert-file argument.

I tested with:
- a non-existing certificate file
- an invalid certificate file
- a valid certificate file without the necessary certificates
- a certificate file with just the bigtable.googleapis.com cert chain
- the default ssl certs file on my workstation
- The required valid/certificates supplied in our internal test environment

BONUS bugfix:
Don't log the gcloud default project unless we're actually going to use
it (i.e. don't print it if -project was already specified).

Change-Id: Ice59b5875c9a816f7ea9a8e27617539e14a61b62
Reviewed-on: https://code-review.googlesource.com/10090
Reviewed-by: Igor Bernstein <igorbernstein@google.com>