trace: always use host to name spans
URLs without any path are uploaded with an empty name.
Stackdriver trace list doesn't render traces without a name,
and we drop information while omiting the host name.

When naming spans, we are dropping the scheme and query
params because they are not statically interesting and can
be retrieved from the label.

Fixes #514.

Change-Id: I84681c7624cb6dd088bef731bf37c3fab472b257
Reviewed-on: https://code-review.googlesource.com/11072
Reviewed-by: John Dethridge <jcd@golang.org>