spanner: fix time.Time comparisons for upcoming Go1.9 monotonic times
An upcoming change in Go1.9 changes the time.Time type to make a
measurement of both the wall clock and a monotonic clock. The
existence of a monotonic measurement makes it such that direct
equality and use of reflect.DeepEqual now fail. Fix by using a
specific date instead of time.Now, so that equality continues to
work. Do note that the documentation for time.Time has alwys
discouraged the use of == and instead recommended the use of the Equal
method.

See https://golang.org/design/12914-monotonic

Change-Id: I733678c5552e1618751b7c41d444230666def381
Reviewed-on: https://code-review.googlesource.com/11254
Reviewed-by: Damien Neil <dneil@google.com>
Reviewed-by: Vikas Kedia <vikask@google.com>