pubsub: check early if streaming iterator is already drained
Fix the fake streaming server to behave more realistically, by waiting
for a CloseSend from the client before ending the stream.

This revealed that the streaming iterator has an unnecessary delay,
due to the stop method blocking on the channel when there were no more
messages. So we now check that case before reading from the channel.

Also, s/Skip/SkipNow/ in tests.

Change-Id: Iacd826c07f6d327475c69e5002cab2381e47d9a5
Reviewed-on: https://code-review.googlesource.com/11374
Reviewed-by: kokoro <noreply+kokoro@google.com>
Reviewed-by: Michael Darakananda <pongad@google.com>
Reviewed-by: Kamal Aboul-Hosn <aboulhosn@google.com>