pubsub: deflake end-to-end test
This test failed occasionally because messages were en route when the
RPC was canceled. Such messages are not redelivered for quite a while,
so the test incorrectly concluded it had quiesced.

I extended the time that had to elapse before declaring quiescence,
and timed it from the last received message instead of using a fixed
interval.

I also left in some useful log messages that helped with debugging,
and may help again.

This test still occasionally fails, because messages get
redelivered. This fix doesn't address that issue.

Change-Id: Icb885c979c61e320a29f1c59ec5ddb85f9eec55f
Reviewed-on: https://code-review.googlesource.com/11655
Reviewed-by: kokoro <noreply+kokoro@google.com>
Reviewed-by: Michael Darakananda <pongad@google.com>
Reviewed-by: Jonathan Amsterdam <jba@google.com>