logging: add Logs method; fix delete test
Add the Logs method to logadmin to support the ListLogs RPC.

Rewrite the test that deletes logs to delete only logs that it can see
with Logs. This addresses the flakiness that resulted from calling
DeleteLog on a log that was just created. That flakiness may result
from the fact that the logs are written with a timestamp given by the
server, but our delete request is with a timestamp from the client's
local machine. There is time skew between the two clocks sometimes,
such that the "delete-after" timestamp given by the client is always
after the timestamp of the logs on the server. So they are never
deleted.

Change-Id: I999c370906b0bd859f1356118087bca0ba70e8db
Reviewed-on: https://code-review.googlesource.com/11871
Reviewed-by: kokoro <noreply+kokoro@google.com>
Reviewed-by: Sarah Adams <shadams@google.com>