spanner: improved KeySet design
BREAKING CHANGE.

Make KeySet an interface. That way, an individual Key or KeyRange can
act as KeySet directly.

Remove the Range function for turning a KeyRange into a KeySet, since
a KeyRange is already a KeySet.

Remove the Keys function for combining a multiple Keys into a KeySet; its
job can be done by the function that builds a union of KeySets.

Since it will probably be used a lot more, rename UnionKeySets to KeySets.

Generalize Delete (for constructing a delete mutation) to take a KeySet
instead of a Key, and remove DeleteKeyRange.

Other changes:

Support a nil KeySet in a delete mutation.

Move PrefixRange to a method on Key (saves the user typing "spanner.")
and rename it AsPrefix.

Change-Id: I0dd1076cb11c2612daf4668a514050a17f958f23
Reviewed-on: https://code-review.googlesource.com/13230
Reviewed-by: Vikas Kedia <vikask@google.com>