pubsub: fix fake server StreamingPull (again)
The former fix didn't work when the server StreamingPull tried to
return an error, because it didn't let the method return.

Fix by observing that there is nothing special about the initial
Recv. Once all Recvs happen in a goroutine, we can return to using
wg.Add/Done as before.

Also, limit the fake test with errors to one goroutine, because
the fake server returns only a single error before hanging.

Change-Id: I84cf2ef58f5ffe0ecb311630ee8b270099b665b8
Reviewed-on: https://code-review.googlesource.com/14670
Reviewed-by: kokoro <noreply+kokoro@google.com>
Reviewed-by: Michael Darakananda <pongad@google.com>