pubsub: remove receiver channel
Remove the channel that mediated between the stream Recv and the
iterator's Next method. Subscription.Receive can more-or-less directly
call the stream Recv. This simplifies the logic of the (former)
iterator.

This is the first change in a series that will involve renaming (since
streamingPullIterator is not an iterator anymore) and probably other
simplifications.

Change-Id: Ia3308310a0e30e401dc9dae9499a2e226837e403
Reviewed-on: https://code-review.googlesource.com/15310
Reviewed-by: kokoro <noreply+kokoro@google.com>
Reviewed-by: Michael Darakananda <pongad@google.com>