bigquery: export ProjectID on DatasetIterator
Allow users to set the project ID of a DatasetIterator by assigning to a
field before use, as we do for other properties of the iterator.

Deprecate the DatasetsInProject method.

Can anyone think of a reason we didn't do it this way in the first place?
It seems simpler and more consistent.

Also, document that the iterator's exported fields should be set initially.

Change-Id: Ic72cc88c125599a20e085caa0c6f6296eb31f21d
Reviewed-on: https://code-review.googlesource.com/16670
Reviewed-by: kokoro <noreply+kokoro@google.com>
Reviewed-by: Sai Cheemalapati <saicheems@google.com>