bigquery: remove job ID configuration from job config structs
Factor out the fields for configuring job ID into a separate struct.
Move it out of the XXXConfig structs and into the Loader/Copier/etc.,
since it's not part of the job config.

This is prep for populating the config types from fetched job info.

Technically this is a breaking change, but only if using XXXConfig
literals. This is unlikely, since they only occur embedded in other
structs which must be created via a function call.

Change-Id: Ia7cd3029e476929ed0d977045fde04b1f75c9c97
Reviewed-on: https://code-review.googlesource.com/18691
Reviewed-by: Michael Darakananda <pongad@google.com>