firestore: remove merge option that splits on dots.
BREAKING CHANGE: Merge is removed; MergePaths is renamed Merge.

The original Merge SetOption took a list of strings and split them on
dots. This could cause confusion, because the Set method does not split
its data's keys on dots.

For example, in

  Set({"a.b": 1, "a": {"b": 2}}, Merge("a.b"))

it appears that the path in the merge option refers to the field with
value 1, but it actually refers to the field with value 2.

It is better to eliminate such confusion by requiring that the merge
fields be expressed as FieldPaths.

Change-Id: Ia8a7e662ebbac5a3bab4efb15a359170fa0dc14c
Reviewed-on: https://code-review.googlesource.com/19252
Reviewed-by: kokoro <noreply+kokoro@google.com>
Reviewed-by: Michael Darakananda <pongad@google.com>