spanner: add OpenCensus spans to some top-level methods
Create new child spans around most method calls.

This will not capture decoding when reading rows. I'm pretty sure that
the StreamingRead span is only for the initial call, not each message
receive. The problem with using spans or annotations to get
information about those is that they are interleaved
(receive-decode-handToCaller-receive-decode-handToCaller...) so we'd
end up with potentially millions of spans or annotations.

I tested with this sample program: https://gist.github.com/jba/278f7bda002781702c171b8d1ca0e679. It produced the output

     0  > main
     0      > cloud.google.com/go/spanner.NewClient
     0      < cloud.google.com/go/spanner.NewClient (0.0s)
     0      > cloud.google.com/go/spanner.Apply
     0          > Sent.google.spanner.v1.Spanner.CreateSession
  1430          < Sent.google.spanner.v1.Spanner.CreateSession (1.4s)
  1430          > Sent.google.spanner.v1.Spanner.Commit
  1612          < Sent.google.spanner.v1.Spanner.Commit (0.2s)
  1612      < cloud.google.com/go/spanner.Apply (1.6s)
  1612      > cloud.google.com/go/spanner.Read
  1612          > Sent.google.spanner.v1.Spanner.StreamingRead
  1705          < Sent.google.spanner.v1.Spanner.StreamingRead (0.1s)
  1612      < cloud.google.com/go/spanner.Read (0.0s)
  1880  < main (1.9s)

I'm not sure of the best way to educate users about this new
functionality.

Change-Id: If60b858c3b855f21f86abe0e2409c8657c3d9741
Reviewed-on: https://code-review.googlesource.com/20550
Reviewed-by: Vikas Kedia <vikask@google.com>