firestore: handle canceled contexts uniformly
Verify that canceling the context passed to a watch causes the watch
to stop immediately.

Also, return a Status with codes.Canceled rather than a
context.Canceled error when sleeping during backoff, so that there is
no distinction between canceling the context during an RPC or while
sleeping between RPCs.

Change-Id: Ic5fbcd881b8cc17e20fdeb7cb2625cdfe9c33c37
Reviewed-on: https://code-review.googlesource.com/25870
Reviewed-by: Jean de Klerk <deklerk@google.com>