pubsub: use unary RPCs for acks and modacks
Send acks and modacks via the unary RPCs instead of over the
stream. Ping the stream periodically while we are processing messages
by sending an empty message.

These changes help mitigate the effects of buffering on streaming
pull. If the client doesn't process messages quickly, messages sent by
the service will get buffered along the network, such as in the gRPC
layer. The stream close signal is also subject to buffering, since it
is sent in band with messages.

One problem with buffering is that the client may be sending acks and
modacks on a stream that has been closed. It can't know this, however,
until it reads enough messages to see the stream close signal. To make
sure acks and modacks are processed even if the stream has been
closed, this CL changes the client to use the non-stream RPCs
Acknowledge and ModifyAckDeadline instead.

This creates another problem. If there is no stream traffic for more
than one minute, the stream will close due to idleness, and the
expiration timer for messages that are still in the buffer will
start. If the client is slow, these messages will expire before the
client reads them. To ameliorate this problem, the client sends an
empty request to the stream every 30 seconds while messages are being
processed. This ping keeps the stream alive and prevents buffered
messages from expiring too soon.

This CL also adds some useful debugging code that can be enabled
by setting a build tag.

Change-Id: I72a00e887ffc2e32ded1e394bbcb4b005b15f90c
Reviewed-on: https://code-review.googlesource.com/27350
Reviewed-by: Michael Darakananda <pongad@google.com>