httpreplay: use []byte for initial instead of interface{}
Originally, a Recorder would take an interface{} for initial state,
call json.Marshal on it, and then the Replayer would call
json.Unmarshal to get it back.

That doesn't work, because json.Marshal/Unmarshal do not round-trip in
general.  You need to pass the same type to Unmarshal that you passed
to Marshal, and the Recorder was passing interface{}.

Example: you use a time.Time for the initial state in the recorder. It
marshals as a string. When you unmarshal that into an interface{}, you
get a string, not a time.Time.

Changed the initial state to []byte. The user is responsible for
serialization. This is what rpcreplay does.

Change-Id: I59d67ca25a3b8ca812085feee481f7c150eeed88
Reviewed-on: https://code-review.googlesource.com/28010
Reviewed-by: Jean de Klerk <deklerk@google.com>