rpcreplay: Implement callbacks before read/write
There is a good case for users to be able to scrub sensitive information
from any RPCs that are saved, particularly if the replay files are in
source control. This change allows users to intercept messages before
they are written to the replay file and before they are searched for in
the replay file.

Change-Id: Ied316c5179ceaed5822a1ef1f607a3e00811fa6c
Reviewed-on: https://code-review.googlesource.com/28670
Reviewed-by: kokoro <noreply+kokoro@google.com>
Reviewed-by: Jonathan Amsterdam <jba@google.com>