pubsub: support labels and geofencing
This CL supports topic MessageStoragePolicy (aka "geofencing") as well
as labels on topics and subscriptions.

- Subscription: Create with, get and update labels.
- Topic:
  - Get and update labels (but not create, because it would be a breaking change).
  - Get and "update" MessageStoragePolicy.

MessageStoragePolicy is unusual. You can't set or update it because
it's configured at the project or organization level. In order to observe
it in the topic resource, you have to call UpdateTopic with
"message_storage_policy" in the field mask (but, again, you must *not*
populate the actual field).

The CL is missing an integration test for MessageStoragePolicy,
because that requires configuring a new project in a special testing
org. A CL with the test will be forthcoming.

Change-Id: I0c72cc4c71520a299695acc21bfd3d477bba9eba
Reviewed-on: https://code-review.googlesource.com/30550
Reviewed-by: Jean de Klerk <deklerk@google.com>
Reviewed-by: kokoro <noreply+kokoro@google.com>